+++ source/ProcessManager.Client/Processes/BRS_023_027/V1/INotifyAggregatedMeasureDataClientV1.cs
@@ -33,14 +33,14 @@ public interface INotifyAggregatedMeasureDataClientV1
     /// <summary>
     /// Get information for BRS-023 or BRS-027 calculation orchestration instance.
     /// </summary>
-    public Task<NotifyAggregatedMeasureDataDtoV1> GetCalculationOrchestrationInstanceAsync(
+    public Task<OrchestrationInstanceDto<NotifyAggregatedMeasureDataInputV1>> GetCalculationOrchestrationInstanceAsync(
         Guid id,
         CancellationToken cancellationToken);
 
     /// <summary>
     /// Get all BRS-023 or BRS-027 calculation orchestration instances filtered by given parameters.
     /// </summary>
-    public Task<IReadOnlyCollection<NotifyAggregatedMeasureDataDtoV1>> SearchCalculationOrchestrationInstancesAsync(
+    public Task<IReadOnlyCollection<OrchestrationInstanceDto<NotifyAggregatedMeasureDataInputV1>>> SearchCalculationOrchestrationInstancesAsync(
         OrchestrationInstanceLifecycleStates? lifecycleState,
         OrchestrationInstanceTerminationStates? terminationState,
         IReadOnlyCollection<CalculationTypes>? calculationTypes,
