+++ source/ProcessManager/ProcessManager.csproj
@@ -33,6 +33,7 @@
     <Folder Include="Api\Processes\BRS_023_027\V1\Model\" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\Shared\ProcessManager\Api\Model\OrchestrationInstanceTypedDto.cs" Link="Api\Model\OrchestrationInstanceTypedDto.cs" />
     <Compile Include="..\Shared\ProcessManager\Api\Model\OrchestrationInstance\OrchestrationInstanceDto.cs" Link="Api\Model\OrchestrationInstance\OrchestrationInstanceDto.cs" />
     <Compile Include="..\Shared\ProcessManager\Api\Model\OrchestrationInstance\OrchestrationInstanceLifecycleStates.cs" Link="Api\Model\OrchestrationInstance\OrchestrationInstanceLifecycleStates.cs" />
     <Compile Include="..\Shared\ProcessManager\Api\Model\OrchestrationInstance\OrchestrationInstanceLifecycleStatesDto.cs" Link="Api\Model\OrchestrationInstance\OrchestrationInstanceLifecycleStatesDto.cs" />
@@ -40,7 +41,6 @@
     <Compile Include="..\Shared\ProcessManager\Api\Model\OrchestrationInstance\OrchestrationStepLifecycleStateDto.cs" Link="Api\Model\OrchestrationInstance\OrchestrationStepLifecycleStateDto.cs" />
     <Compile Include="..\Shared\ProcessManager\Api\Model\OrchestrationInstance\OrchestrationStepLifecycleStates.cs" Link="Api\Model\OrchestrationInstance\OrchestrationStepLifecycleStates.cs" />
     <Compile Include="..\Shared\ProcessManager\Api\Model\ScheduleOrchestrationInstanceDto.cs" Link="Api\Model\ScheduleOrchestrationInstanceDto.cs" />
-    <Compile Include="..\Shared\ProcessManager\Api\Processes\BRS_023_027\V1\Model\NotifyAggregatedMeasureDataDtoV1.cs" Link="Api\Processes\BRS_023_027\V1\Model\NotifyAggregatedMeasureDataDtoV1.cs" />
     <Compile Include="..\Shared\ProcessManager\Orchestrations\Processes\BRS_023_027\V1\Model\CalculationTypes.cs" Link="Api\Processes\BRS_023_027\V1\Model\CalculationTypes.cs" />
     <Compile Include="..\Shared\ProcessManager\Orchestrations\Processes\BRS_023_027\V1\Model\NotifyAggregatedMeasureDataInputV1.cs" Link="Api\Processes\BRS_023_027\V1\Model\NotifyAggregatedMeasureDataInputV1.cs" />
   </ItemGroup>
